requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pillow==10.4.0
numpy==1.26.4
faiss-cpu==1.9.0.post1
torch==2.2.2
transformers==4.41.2
sentence-transformers==2.2.2
requests==2.32.5